+++ client-app/src/admin/AppModel.ts
@@ -1,5 +1,5 @@
 import {XH} from '@xh/hoist/core';
-import {AppModel as BaseAppModel} from '@xh/hoist/admin/AppModel';
+import {AppModel as HoistAdminAppModel} from '@xh/hoist/admin/AppModel';
 import {Icon} from '@xh/hoist/icon';
 import {PortfolioService} from '../core/svc/PortfolioService';
 import {roadmapTab} from './roadmap/RoadmapTab';
@@ -7,17 +7,20 @@ import {testsTab} from './tests/TestsTab';
 import {wipTab} from './wip/WipTab';
 import {OauthService} from '../core/svc/OauthService';
 
-export class AppModel extends BaseAppModel {
+export class AppModel extends HoistAdminAppModel {
     static override instance: AppModel;
 
+    static override async preAuthAsync() {
+        await XH.installServicesAsync(OauthService);
+    }
+
     override async initAsync() {
+        await super.initAsync();
         await XH.installServicesAsync(PortfolioService);
     }
 
-    static override async preAuthAsync() {
-        if (XH.appSpec.isSSO) {
-            await XH.installServicesAsync(OauthService);
-        }
+    override async logoutAsync() {
+        XH.oauthService.enabled ? await XH.oauthService.logoutAsync() : await super.logoutAsync();
     }
 
     //------------------------