+++ client-app/src/core/svc/OauthService.ts
@@ -2,7 +2,6 @@ import createAuth0Client, {Auth0Client} from '@auth0/auth0-spa-js';
 import {HoistService, PlainObject, XH} from '@xh/hoist/core';
 import {never, wait} from '@xh/hoist/promise';
 import {SECONDS} from '@xh/hoist/utils/datetime';
-import {logError, logInfo, logWarn} from '@xh/hoist/utils/js';
 
 /**
  * Coordinates OAuth-based login for the user-facing desktop and mobile apps.
@@ -26,6 +25,12 @@ import {logError, logInfo, logWarn} from '@xh/hoist/utils/js';
 export class OauthService extends HoistService {
     static instance: OauthService;
 
+    /**
+     * Is OAuth enabled in this application?  For bootstrapping, troubleshooting
+     * and mobile development, we allow running in a non-SSO mode.
+     */
+    enabled: boolean;
+
     auth0: Auth0Client;
     /** Authenticated user info as provided by Auth0. */
     user: PlainObject;
@@ -43,7 +48,8 @@ export class OauthService extends HoistService {
             url: 'oauthConfig'
         }).catchDefault());
 
-        if (config?.isEnabled === false) {
+        this.enabled = config.enabled;
+        if (!this.enabled) {
             XH.appSpec.isSSO = false;
             return;
         }
@@ -51,7 +57,8 @@ export class OauthService extends HoistService {
         if (!config?.domain || !config?.clientId) {
             throw XH.exception(`
                 Unable to init OAuthService - expected config not returned by server. 
-                Please log on as a local admin user and review this instance's soft configuration entries.
+                Please review the settings in configs "auth0ClientId" and "auth0Domain".
+                Default values for these configs are provided in "Bootstrap.groovy"
             `);
         }
 
@@ -77,14 +84,14 @@ export class OauthService extends HoistService {
                 const cleanUrl = window.location.toString().replace(qString, '');
                 window.history.replaceState({}, document.title, cleanUrl);
             } catch (e) {
-                logWarn([`Caught while attempting to get redirectResults`, e], this);
+                this.logWarn(`Caught while attempting to get redirectResults`, e);
             }
         }
 
         if (!isAuthenticated) {
             // If still not authenticated, we are either coming in fresh or were unable to confirm a
             // successful auth via redirect handler. Trigger interactive login.
-            logInfo(`Not authenticated - logging in....`, this);
+            this.logInfo(`Not authenticated - logging in....`);
             await auth0.loginWithRedirect();
             await never();
         } else {
@@ -92,17 +99,20 @@ export class OauthService extends HoistService {
             this.user = await this.auth0.getUser();
             this.idToken = await this.getIdTokenAsync();
 
-            logInfo([`Authenticated OK`, this.user?.email, this.user], this);
+            this.logInfo(`Authenticated OK`, this.user?.email, this.user);
             this.installDefaultFetchServiceHeaders();
         }
     }
 
+    //------------------
+    // Implementation
+    //-----------------
     private async getIdTokenAsync() {
         const claims = await this.auth0.getIdTokenClaims();
         return claims?.__raw;
     }
 
-    private async checkAuthAsync() {
+    private async checkAuthAsync(): Promise<boolean> {
         return this.auth0.isAuthenticated();
     }
 
@@ -110,6 +120,7 @@ export class OauthService extends HoistService {
      * Logout of both Hoist session and Auth0 Oauth session (if active).
      */
     async logoutAsync() {
+        if (!this.enabled) return;
         try {
             const hasOauth = await this.checkAuthAsync();
             if (hasOauth) {
@@ -124,7 +135,7 @@ export class OauthService extends HoistService {
                 await wait(10 * SECONDS);
             }
         } catch (e) {
-            logError(['Error during logout request', e], this);
+            this.logError('Error during logout request', e);
         }
     }