+++ grails-app/services/io/xh/toolbox/security/Auth0Service.groovy
@@ -2,11 +2,13 @@ package io.xh.toolbox.security
 
 import io.xh.hoist.BaseService
 import io.xh.hoist.config.ConfigService
+import io.xh.hoist.http.JSONClient
 import io.xh.hoist.json.JSONParser
-import io.xh.hoist.json.JSONSerializer
+import org.apache.hc.client5.http.classic.methods.HttpGet
 import org.jose4j.jwk.JsonWebKeySet
 import org.jose4j.jwk.VerificationJwkSelector
 import org.jose4j.jws.JsonWebSignature
+
 
 import static io.xh.hoist.util.InstanceConfigUtils.getInstanceConfig
 
@@ -15,17 +17,20 @@ import static io.xh.hoist.util.InstanceConfigUtils.getInstanceConfig
  */
 class Auth0Service extends BaseService {
 
-    static clearCachesConfigs = ['auth0Domain', 'auth0ClientId', 'auth0Jwks']
+    private JSONClient _jsonClient
+
+    static clearCachesConfigs = ['auth0Domain', 'auth0ClientId']
 
     ConfigService configService
 
     Map getClientConfig() {
 
         if (getInstanceConfig('useOAuth') == 'false') {
-            return [isEnabled: false]
+            return [enabled: false]
         }
 
         return [
+            enabled: true,
             clientId: clientId,
             domain: domain
         ]
@@ -67,21 +72,26 @@ class Auth0Service extends BaseService {
     //------------------------
     // Implementation
     //------------------------
-    private JsonWebKeySet _jkws
-    JsonWebKeySet getJsonWebKeySet() {
-        if (!_jkws) {
-            // Yes, this is a bit odd - we get a JSON config then re-serialize to JSON to pass to
-            // JWKS ctor. Could store as string config, but keeping as JSON in configService gives
-            // us a nicer UI to view/update in admin console.
-            def jwksJson = JSONSerializer.serialize(configService.getMap('auth0Jwks'))
-            _jkws = new JsonWebKeySet(jwksJson)
+    private JSONClient getClient() {
+        if (!_jsonClient) {
+            _jsonClient = new JSONClient()
+        }
+        return _jsonClient
+    }
 
-            if (!_jkws.jsonWebKeys.size()) {
-                throw new RuntimeException("Unable to build valid key set from 'auth0Jwks' app config")
+    private JsonWebKeySet _jwks
+    JsonWebKeySet getJsonWebKeySet() {
+        if (!_jwks) {
+            def url = "https://${domain}/.well-known/jwks.json",
+                jwksJson = client.executeAsString(new HttpGet(url))
+            _jwks = new JsonWebKeySet(jwksJson)
+
+            if (!_jwks.jsonWebKeys.size()) {
+                throw new RuntimeException("Unable to build valid key set from remote JWKS endpoint.")
             }
         }
 
-        return _jkws
+        return _jwks
     }
 
     private String getClientId() {
@@ -93,8 +103,9 @@ class Auth0Service extends BaseService {
     }
 
     void clearCaches() {
-        _jkws = null
         super.clearCaches()
+        _jsonClient = null
+        _jwks = null
     }
 
     Map getAdminStats() {[